+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 **FaaS**
 - Add `pull` and `tag` parameters to `build` method.
 - Add `gateway` and `namespace` parameters to `deploy` and `invoke` methods.
+- `gateway` and `user` parameters no longer mandatory for `login` method.
 
 **GoSS**
 - Add `logLevel` parameter to `server` and `validate` methods.